ai-service/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
numpy==1.26.4
pydantic==2.7.4
python-dotenv==1.0.1
google-generativeai==0.8.4
